+++ assets/javascripts/kitten/components/form/text-input/index.js
@@ -157,6 +157,17 @@ const StyledInput = styled.input`
   @media (max-width: ${ScreenConfig.XS.max}px) {
     font-size: ${stepToRem(0)};
   }
+
+  // DARK BACKGROUND
+  &.k-Form-TextInput--darkBackground {
+    background-color: ${COLORS.grey1};
+    border-color: ${COLORS.grey1};
+    color: ${COLORS.background1};
+
+    ::placeholder {
+      color: ${COLORS.line2};
+    }
+  }
 `
 
 const StyledTextareaContainer = styled.div`
@@ -221,6 +232,16 @@ const StyledTextareaContainer = styled.div`
     border-radius: var(--input-height);
   }
 
+  &.k-Form-TextInput--darkBackground {
+    background-color: ${COLORS.grey1};
+    border-color: ${COLORS.grey1};
+    color: ${COLORS.background1};
+
+    ::placeholder {
+      color: ${COLORS.line2};
+    }
+  }
+
   .k-Form-TextInput__textareaGradient {
     position: absolute;
     left: ${pxToRem(10)};
@@ -254,6 +275,7 @@ export class TextInput extends PureComponent {
     digits: PropTypes.number,
     variant: PropTypes.oneOf(['orion', 'andromeda']),
     rounded: PropTypes.bool,
+    darkBackground: PropTypes.bool,
   }
 
   static defaultProps = {
@@ -267,6 +289,7 @@ export class TextInput extends PureComponent {
     digits: null,
     variant: 'orion',
     rounded: false,
+    darkBackground: false,
   }
 
   render() {
@@ -283,6 +306,7 @@ export class TextInput extends PureComponent {
       className,
       style,
       rounded,
+      darkBackground,
       ...others
     } = this.props
 
@@ -312,6 +336,7 @@ export class TextInput extends PureComponent {
                 'k-Form-TextInput--disabled': disabled,
                 'k-Form-TextInput--alignCenter': center,
                 'k-Form-TextInput--rounded': rounded,
+                'k-Form-TextInput--darkBackground': darkBackground,
               },
             )}
             style={
@@ -340,6 +365,7 @@ export class TextInput extends PureComponent {
               'k-Form-TextInput--disabled': disabled,
               'k-Form-TextInput--alignCenter': center,
               'k-Form-TextInput--rounded': rounded,
+              'k-Form-TextInput--darkBackground': darkBackground,
             },
           )}
           style={